+++ src/Helper/XMLParser.php
@@ -3,7 +3,7 @@
 namespace PhpXmlRpc\Helper;
 
 use PhpXmlRpc\PhpXmlRpc;
-use PhpXmlRpc\Traits\LoggerAware;
+use PhpXmlRpc\Traits\DeprecationLogger;
 use PhpXmlRpc\Value;
 
 /**
@@ -20,7 +20,7 @@ use PhpXmlRpc\Value;
  */
 class XMLParser
 {
-    use LoggerAware;
+    use DeprecationLogger;
 
     const RETURN_XMLRPCVALS = 'xmlrpcvals';
     const RETURN_EPIVALS = 'epivals';
@@ -1012,14 +1012,14 @@ class XMLParser
 
     public function __set($name, $value)
     {
-        //trigger_error('setting property Response::' . $name . ' is deprecated', E_USER_DEPRECATED);
-
         switch ($name) {
             // this should only ever be called by subclasses which overtook `parse()`
             case 'accept':
+                $this->logDeprecation('Setting property XMLParser::' . $name . ' is deprecated');
                 $this->current_parsing_options['accept'] = $value;
                 break;
             default:
+                /// @todo throw instead? There are very few other places where the lib trigger errors which can potentially reach stdout...
                 $trace = debug_backtrace(DEBUG_BACKTRACE_IGNORE_ARGS);
                 trigger_error('Undefined property via __set(): ' . $name . ' in ' . $trace[0]['file'] . ' on line ' . $trace[0]['line'], E_USER_WARNING);
         }
@@ -1027,10 +1027,9 @@ class XMLParser
 
     public function __isset($name)
     {
-        //trigger_error('checking property Response::' . $name . ' is deprecated', E_USER_DEPRECATED);
-
         switch ($name) {
             case 'accept':
+                $this->logDeprecation('Checking property XMLParser::' . $name . ' is deprecated');
                 return isset($this->current_parsing_options['accept']);
             default:
                 return false;
@@ -1042,9 +1041,11 @@ class XMLParser
         switch ($name) {
             // q: does this make sense at all?
             case 'accept':
+                $this->logDeprecation('Unsetting property XMLParser::' . $name . ' is deprecated');
                 unset($this->current_parsing_options['accept']);
                 break;
             default:
+                /// @todo throw instead? There are very few other places where the lib trigger errors which can potentially reach stdout...
                 $trace = debug_backtrace(DEBUG_BACKTRACE_IGNORE_ARGS);
                 trigger_error('Undefined property via __unset(): ' . $name . ' in ' . $trace[0]['file'] . ' on line ' . $trace[0]['line'], E_USER_WARNING);
         }