+++ src/Server.php
@@ -9,7 +9,7 @@ use PhpXmlRpc\Helper\Interop;
 use PhpXmlRpc\Helper\Logger;
 use PhpXmlRpc\Helper\XMLParser;
 use PhpXmlRpc\Traits\CharsetEncoderAware;
-use PhpXmlRpc\Traits\LoggerAware;
+use PhpXmlRpc\Traits\DeprecationLogger;
 use PhpXmlRpc\Traits\ParserAware;
 
 /**
@@ -18,7 +18,7 @@ use PhpXmlRpc\Traits\ParserAware;
 class Server
 {
     use CharsetEncoderAware;
-    use LoggerAware;
+    use DeprecationLogger;
     use ParserAware;
 
     const OPT_ACCEPTED_COMPRESSION = 'accepted_compression';